feat: added imports to index.ts refactor: restructured chevron-down.tsx add react.Memo
