+++ src/assets/icons/ChevronDown.tsx
@@ -1,9 +1,10 @@
-import React from 'react'
+import { Ref, SVGProps, forwardRef, memo } from 'react'
 
-export const ChevronDown = (props: React.SVGProps<SVGSVGElement>) => (
+const SvgChevronDown = (props: SVGProps<SVGSVGElement>, ref: Ref<SVGSVGElement>) => (
   <svg
     fill={'none'}
     height={'24'}
+    ref={ref}
     viewBox={'0 0 24 24'}
     width={'24'}
     xmlns={'http://www.w3.org/2000/svg'}
@@ -24,3 +25,7 @@ export const ChevronDown = (props: React.SVGProps<SVGSVGElement>) => (
     </defs>
   </svg>
 )
+const ForwardRef = forwardRef(SvgChevronDown)
+const Memo = memo(ForwardRef)
+
+export default Memo
