+++ src/assets/icons/index.ts
@@ -1,0 +1,3 @@
+export { default as ChevronDown } from './ChevronDown'
+export { default as FlagRussia } from './FlagRussia'
+export { default as FlagUnitedKingdom } from './FlagUnitedKingdom'
