+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * from '@/components/button/button'
+export * from '@/components/select/select'
